Fix compilation logic, and only forward search on sumatra when compilation is done
