+++ src/nl/hannahsten/texifyidea/run/OpenCustomPdfViewerListener.kt
@@ -4,12 +4,13 @@ import com.intellij.execution.process.ProcessEvent
 import com.intellij.execution.process.ProcessListener
 import com.intellij.openapi.util.Key
 import nl.hannahsten.texifyidea.TeXception
+import nl.hannahsten.texifyidea.run.latex.LatexRunConfiguration
 import java.io.IOException
 
 /**
  * Run the given command after the process has succeeded.
  */
-class OpenPdfViewerListener(val command: Array<String>, val failSilently: Boolean = false) : ProcessListener {
+class OpenCustomPdfViewerListener(val command: Array<String>, val failSilently: Boolean = false, val runConfig: LatexRunConfiguration) : ProcessListener {
     override fun processTerminated(event: ProcessEvent) {
         if (event.exitCode == 0) {
             try {
@@ -21,6 +22,9 @@ class OpenPdfViewerListener(val command: Array<String>, val failSilently: Boolea
                 }
             }
         }
+
+        // Reset to default
+        runConfig.allowFocusChange = true
     }
 
     override fun processWillTerminate(event: ProcessEvent, willBeDestroyed: Boolean) {
