+++ src/nl/hannahsten/texifyidea/run/linuxpdfviewer/evince/EvinceConversation.kt
@@ -70,7 +70,7 @@ object EvinceConversation : ViewerConversation() {
         else {
             // If the user used the forward search menu action
             if (pdfPath == null) {
-                Notification("EvinceConversation", "Could not execute forward search", "Please make sure you have compiled the document first, or if you are using autocompile execute forward search twice.", NotificationType.ERROR).notify(project)
+                Notification("EvinceConversation", "Could not execute forward search", "Please make sure you have compiled the document first.", NotificationType.ERROR).notify(project)
             }
             else {
                 throw TeXception("Could not execute forward search with Evince because something went wrong when finding the pdf file at $pdfPath")
