+++ src/nl/hannahsten/texifyidea/run/sumatra/SumatraForwardSearchListener.kt
@@ -1,7 +1,9 @@
 package nl.hannahsten.texifyidea.run.sumatra
 
-import com.intellij.execution.process.ProcessHandler
+import com.intellij.execution.process.ProcessEvent
+import com.intellij.execution.process.ProcessListener
 import com.intellij.execution.runners.ExecutionEnvironment
+import com.intellij.openapi.util.Key
 import nl.hannahsten.texifyidea.TeXception
 import nl.hannahsten.texifyidea.psi.LatexEnvironment
 import nl.hannahsten.texifyidea.run.latex.LatexRunConfiguration
@@ -15,15 +17,22 @@ import nl.hannahsten.texifyidea.util.parentOfType
 import org.jetbrains.concurrency.runAsync
 
 /**
- * Provides forward search for SumatraPDF.
+ * @author Sten Wessel
  */
-class SumatraForwardSearch {
+class SumatraForwardSearchListener(val runConfig: LatexRunConfiguration,
+                                   private val executionEnvironment: ExecutionEnvironment,
+                                   private val focusAllowed: Boolean
+) : ProcessListener {
 
-    /**
-     * Execute forward search based on the given environment.
-     */
-    fun execute(handler: ProcessHandler, runConfig: LatexRunConfiguration, executionEnvironment: ExecutionEnvironment, focusAllowed: Boolean = true) {
-        handler.addProcessListener(OpenSumatraListener(runConfig))
+    override fun processTerminated(event: ProcessEvent) {
+        // First check if the user provided a custom path to SumatraPDF, if not, check if it is installed
+        if (event.exitCode == 0 && (runConfig.sumatraPath != null || isSumatraAvailable)) {
+            try {
+                SumatraConversation.openFile(runConfig.outputFilePath, sumatraPath = runConfig.sumatraPath)
+            }
+            catch (ignored: TeXception) {
+            }
+        }
 
         // Forward search.
         run {
@@ -57,5 +66,20 @@ class SumatraForwardSearch {
                 }
             }
         }
+
+        // Reset to default
+        runConfig.allowFocusChange = true
+    }
+
+    override fun onTextAvailable(p0: ProcessEvent, p1: Key<*>) {
+        // Do nothing.
+    }
+
+    override fun processWillTerminate(p0: ProcessEvent, p1: Boolean) {
+        // Do nothing.
+    }
+
+    override fun startNotified(p0: ProcessEvent) {
+        // Do nothing.
     }
 }
